Rename all files to *.js; invoke with node -r esm
