+++ README.md
@@ -26,7 +26,7 @@ validate your work:
     npm test                   # Test the parser against JSON AST fixtures.
     npm run lint               # Lint the parser code.
 
-    npm run generate:ebnf      # Generate the EBNF from syntax/grammar.mjs.
+    npm run generate:ebnf      # Generate the EBNF from syntax/grammar.js.
     npm run generate:fixtures  # Generate test fixtures (FTL → JSON AST).
 
     npm run build:guide        # Build the HTML version of the Guide.
