+++ lib/combinators.js
@@ -1,6 +1,6 @@
-import Parser from "./parser.mjs";
-import {Success, Failure} from "./result.mjs";
-import {join} from "./mappers.mjs";
+import Parser from "./parser.js";
+import {Success, Failure} from "./result.js";
+import {join} from "./mappers.js";
 
 export function defer(fn) {
     // Parsers may be defined as defer(() => parser) to avoid cyclic
