+++ lib/ebnf.js
@@ -1,11 +1,11 @@
-import babylon from "babylon";
-import walk from "./walker.mjs";
-import visitor from "./visitor.mjs";
-import serialize from "./serializer.mjs";
+import {parse} from "babylon";
+import walk from "./walker.js";
+import visitor from "./visitor.js";
+import serialize from "./serializer.js";
 
 export default
 function ebnf(source, min_name_length = 0) {
-    let grammar_ast = babylon.parse(source, {sourceType: "module"});
+    let grammar_ast = parse(source, {sourceType: "module"});
     let rules = walk(grammar_ast, visitor);
     let state = {
         max_name_length: Math.max(
@@ -16,4 +16,3 @@ function ebnf(source, min_name_length = 0) {
         .map(rule => serialize(rule, state))
         .join("");
 }
-
