+++ lib/mappers.js
@@ -1,4 +1,4 @@
-import {Abstract} from "./parser.mjs";
+import {Abstract} from "./parser.js";
 
 // Flatten a list up to a given depth.
 // This is useful when a parser uses nested sequences and repeats.
