+++ lib/parser.js
@@ -1,5 +1,5 @@
-import Stream from "./stream.mjs";
-import {Failure} from "./result.mjs";
+import Stream from "./stream.js";
+import {Failure} from "./result.js";
 
 export class Abstract {
     constructor(value) {
