+++ lib/stream.js
