+++ lib/visitor.js
