+++ lib/walker.js
