+++ package.json
@@ -4,19 +4,19 @@
   "version": "1.0.0",
   "private": true,
   "scripts": {
-    "bench": "node --experimental-modules --harmony-async-iteration ./test/bench.mjs ./test/benchmarks/gecko_strings.ftl",
+    "bench": "node -r esm --harmony-async-iteration ./test/bench.js ./test/benchmarks/gecko_strings.ftl",
     "build:guide": "gitbook build guide build/guide",
     "build": "npm run --silent build:guide",
     "ci": "npm run --silent lint && npm test && npm run --silent test:ebnf",
     "clean": "rm -rf build",
     "deploy": "gh-pages -d build",
-    "generate:ebnf": "node --experimental-modules bin/ebnf.mjs ./syntax/grammar.mjs > ./spec/fluent.ebnf",
+    "generate:ebnf": "node -r esm bin/ebnf.js ./syntax/grammar.js > ./spec/fluent.ebnf",
     "generate:fixtures": "make -sC test/fixtures",
     "generate": "npm run --silent generate:ebnf && npm run --silent generate:fixtures",
-    "lint": "eslint **/*.mjs",
-    "test:ebnf": "node --experimental-modules test/ebnf.mjs ./syntax/grammar.mjs ./spec/fluent.ebnf",
-    "test:fixtures": "node --experimental-modules test/parser.mjs ./test/fixtures",
-    "test:unit": "node --experimental-modules test/literals.mjs",
+    "lint": "eslint **/*.js",
+    "test:ebnf": "node -r esm test/ebnf.js ./syntax/grammar.js ./spec/fluent.ebnf",
+    "test:fixtures": "node -r esm test/parser.js ./test/fixtures",
+    "test:unit": "node -r esm test/literals.js",
     "test": "npm run --silent test:fixtures && npm run --silent test:unit"
   },
   "homepage": "https://projectfluent.org",
@@ -41,6 +41,7 @@
     "json-diff": "^0.5.2"
   },
   "dependencies": {
+    "esm": "^3.2.25",
     "minimist": "^1.2.0"
   }
 }
