+++ test/bench.js
@@ -2,17 +2,17 @@ import fs from "fs";
 import perf from "perf_hooks";
 const {PerformanceObserver, performance} = perf;
 
-import FluentSyntax from "fluent-syntax";
-import FluentRuntime from "fluent";
-import {Resource} from "../syntax/grammar.mjs";
-import {readfile} from "./util.mjs";
+import {parse} from "fluent-syntax";
+import {_parse} from "fluent";
+import {Resource} from "../syntax/grammar.js";
+import {readfile} from "./util.js";
 
 let args = process.argv.slice(2);
 
 if (args.length < 1 || 2 < args.length) {
     console.error(
-        "Usage: node --experimental-modules --harmony-async-iteration " +
-        "bench.mjs FTL_FILE [SAMPLE SIZE = 30]");
+        "Usage: node -r esm --harmony-async-iteration " +
+        "bench.js FTL_FILE [SAMPLE SIZE = 30]");
     process.exit(1);
 }
 
@@ -31,8 +31,8 @@ async function main(ftl_file, sample_size = 30) {
 
     let subjects = new Map([
         ["Reference", new Subject("Reference", () => Resource.run(ftl))],
-        ["Tooling", new Subject("Tooling", () => FluentSyntax.parse(ftl))],
-        ["Runtime", new Subject("Runtime", () => FluentRuntime._parse(ftl))],
+        ["Tooling", new Subject("Tooling", () => parse(ftl))],
+        ["Runtime", new Subject("Runtime", () => _parse(ftl))],
     ]);
 
     new PerformanceObserver(items => {
